+++ integration_test/chrome/all_test.exs
@@ -15,3 +15,6 @@ Code.require_file("../cases/browser/send_keys_to_active_element_test.exs", __DIR
 Code.require_file("../cases/browser/window_handles_test.exs", __DIR__)
 Code.require_file("../cases/browser/window_position_test.exs", __DIR__)
 Code.require_file("../chrome/capabilities_test.exs", __DIR__)
+Code.require_file("../cases/feature/use_feature_test.exs", __DIR__)
+Code.require_file("../cases/feature/import_feature_test.exs", __DIR__)
+Code.require_file("../cases/feature/automatic_screenshot_test.exs", __DIR__)