+++ lib/wallaby/session_store.ex
@@ -6,41 +6,68 @@ defmodule Wallaby.SessionStore do
 
   def start_link, do: GenServer.start_link(__MODULE__, :ok, name: __MODULE__)
 
-  def monitor(session), do: GenServer.call(__MODULE__, {:monitor, session}, 10_000)
+  def monitor(session) do
+    GenServer.call(__MODULE__, {:monitor, session}, 10_000)
+  end
 
-  def demonitor(session), do: GenServer.call(__MODULE__, {:demonitor, session})
+  def demonitor(session) do
+    GenServer.call(__MODULE__, {:demonitor, session})
+  end
+
+  def list_sessions_for(owner_pid \\ self()) when is_pid(owner_pid) do
+    :ets.select(:session_store, [{{{:_, :_, :"$1"}, :"$2"}, [{:==, :"$1", owner_pid}], [:"$2"]}])
+  end
 
   def init(:ok) do
     Process.flag(:trap_exit, true)
-    {:ok, %{refs: %{}}}
+    :ets.new(:session_store, [:set, :named_table, :public, read_concurrency: true])
+
+    {:ok, nil}
   end
 
-  def handle_call({:monitor, session}, {pid, _ref}, %{refs: refs} = state) do
+  def handle_call({:monitor, session}, {pid, _ref}, _state) do
     ref = Process.monitor(pid)
-    refs = Map.put(refs, ref, session)
-    {:reply, :ok, %{state | refs: refs}}
+
+    :ets.insert(
+      :session_store,
+      {{ref, session.id, pid}, session}
+    )
+
+    {:reply, :ok, nil}
   end
 
-  def handle_call({:demonitor, session}, _from, %{refs: refs} = state) do
-    case Enum.find(refs, fn {_, value} -> value.id == session.id end) do
-      {ref, _} ->
-        {_, refs} = Map.pop(refs, ref)
-        true = Process.demonitor(ref)
-        {:reply, :ok, %{state | refs: refs}}
+  def handle_call({:demonitor, session}, _from, _state) do
+    result =
+      :ets.select(:session_store, [
+        {{{:"$1", :"$2", :"$3"}, :"$4"}, [{:==, :"$2", session.id}], [{{:"$1", :"$4"}}]}
+      ])
 
-      nil ->
-        {:reply, :ok, state}
+    case result do
+      [{ref, _}] ->
+        true = Process.demonitor(ref)
+        :ets.delete(:session_store, {ref, session.id})
+
+        {:reply, :ok, nil}
+
+      [] ->
+        {:reply, :ok, nil}
     end
   end
 
-  def handle_info({:DOWN, ref, :process, _pid, _reason}, %{refs: refs} = state) do
-    {session, refs} = Map.pop(refs, ref)
+  def handle_info({:DOWN, ref, :process, _pid, _reason}, _state) do
+    [session] =
+      :ets.select(:session_store, [{{{:"$1", :"$2", :"$3"}, :"$4"}, [{:==, :"$1", ref}], [:"$4"]}])
+
     WebdriverClient.delete_session(session)
-    {:noreply, %{state | refs: refs}}
+
+    :ets.delete(:session_store, {ref, session.id})
+
+    {:noreply, nil}
   end
 
-  def terminate(_reason, %{refs: refs}) do
-    Enum.each(refs, fn {_ref, session} -> close_session(session) end)
+  def terminate(_reason, _state) do
+    :ets.tab2list(:session_store)
+    |> Enum.each(fn {_, session} -> close_session(session) end)
   end
 
   defp close_session(session) do